Fix space eating issue
The issue is if there's a less than symbol in the script it comes
through the `ontext` callback separated before and after. So for
example

     if (a < b) { doit(); }

gets split into `if (a ` (space after `a`) and `< b) { doit(); }`
A line in the `ontext` handler scripts trailing whitespec so when
finally reassembled it comes out with

     if (a< b) { doit(); }

Which no longer lints correctly.

I think the solution is to strip the only the last string in each
script tag