fix: scope sharing should not be influenced by globalReturn: true
The 'node' environment sets parserOptions.ecmaFeatures.globalReturn to
'true'.  Because of this, the JS code is considered as wrapped in a
function, so declared variables are not in the global scope.  To keep
the configuration simple, scope sharing is only activated by
parserOptions.sourceType = 'module', so setting globalReturn to true
should still consider variables declared in the script body to be
exported to the following scopes.
Partially fixes #100